pyproject.toml
[project]
name = "mnlp"
version = "0.1.0"
description = ""
authors = [{ name = "Kaan-wq", email = "kaanucar9@gmail.com" }]
requires-python = ">=3.11,<3.15"
readme = "README.md"
dependencies = [
    "torch (>=2.4.0)",
    "wandb (>=0.25.1)",
    "python-dotenv (>=1.2.2)",
    "jupyter (>=1.1.1)",
    "huggingface-hub (>=1.9.0)",
    "transformers (>=5.5.0)",
    "datasets (>=4.8.4)",
    "evaluate (>=0.4.6,<0.5.0)",
    "scipy (>=1.17.1,<2.0.0)",
    "scikit-learn (>=1.8.0,<2.0.0)",
    "accelerate (>=1.13.0,<2.0.0)",
    "sentencepiece (>=0.2.1,<0.3.0)",
    "sacremoses (>=0.1.1,<0.2.0)",
    "sacrebleu (>=2.6.0,<3.0.0)",
    "ipykernel (>=7.2.0,<8.0.0)",
    "gradio-client (>=2.4.0,<3.0.0)",
    "peft (>=0.18.1,<0.19.0)",
    "trl (>=1.0.0,<2.0.0)",
    "bitsandbytes (>=0.49.2,<0.50.0)",
    "codecarbon (>=3.2.6,<4.0.0)",
]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = false

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes — unused imports, undefined names
    "I",   # isort — import sorting
    "UP",  # pyupgrade — modern Python syntax
    "B",   # flake8-bugbear — common bugs
    "C4",   # better comprehensions
    "SIM",  # simplify conditions
    "TCH",  # type-checking imports
]
ignore = [
    "B006",  # mutable default arguments — common in configs
    "SIM108", # ternary operator — sometimes less readable
]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
